Move index and Enum to utils
