+++ python/basix/__init__.py
@@ -7,7 +7,7 @@ from basix import cell, finite_element, lattice, polynomials, quadrature, sobole
 from basix._basixcpp import __version__
 from basix.cell import CellType, geometry, topology
 from basix.finite_element import DPCVariant, ElementFamily, LagrangeVariant, create_custom_element, create_element
-from basix.index import index
+from basix.utils import index
 from basix.interpolation import compute_interpolation_operator
 from basix.lattice import LatticeSimplexMethod, LatticeType, create_lattice
 from basix.maps import MapType
