+++ python/basix/sobolev_spaces.py
@@ -4,7 +4,7 @@ import typing as _typing
 
 from basix._basixcpp import SobolevSpace as _SS
 from basix._basixcpp import sobolev_space_intersection as _ssi
-from basix.enums import Enum
+from basix.utils import Enum
 
 __all__ = ["intersection", "string_to_sobolev_space"]
