+++ python/basix/utils.py
@@ -1,5 +1,8 @@
-"""Enumerations."""
+"""Utility funcitons."""
 
+import typing
+
+from basix._basixcpp import index as _index
 from enum import Enum as _Enum
 
 __all__ = ["Enum"]
@@ -31,3 +34,14 @@ class Enum(_Enum):
         if self.__class__ != other.__class__:
             return NotImplemented
         return self.value >= other.value
+
+
+def index(p: int, q: typing.Optional[int] = None, r: typing.Optional[int] = None) -> int:
+    """TODO."""
+    if q is None:
+        assert r is None
+        return _index(p)
+    elif r is None:
+        return _index(p, q)
+    else:
+        return _index(p, q, r)
